updated excluded extensions for TYPO3 8.7
the 2 authentication classes write a warning log on every FE call as
long as there is no logged in fe_user. this flood the log with warnings
that are no warnings. this behaviour won't change so the only option is
to ignore logs from TYPO3\CMS\Sv\AuthenticationService and
TYPO3\CMS\Core\Authentication\AbstractUserAuthentication. This should be
no problem at all